working, no wear near finished

+++ lib/data_logger/provider/datalog_provider.dart
@@ -135,7 +135,7 @@ class DataLogProvider extends StateNotifier<Uint8List> {
         'tps': tps.toDouble(),
         'afr': modAfr,
         'airboxPressure': modAirBoxPressure,
-        'airTemperature': airTemperature,
+        'airTemperature': airTemperature.toDouble(),
         'coolantTemperature': coolantTemperature.toDouble(),
         'oilPressure': oilPressure,
         'batteryVoltage': modBatteryVoltage,
